Finishing port show / hide
